+++ packages/compass-components/src/index.ts
@@ -125,7 +125,6 @@ export { TabNavBar } from './components/tab-nav-bar';
 export { WorkspaceContainer } from './components/workspace-container';
 export { InlineInfoLink } from './components/inline-info-link';
 export { InteractivePopover } from './components/interactive-popover';
-export { ModalPopover } from './components/modal-popover';
 export { Placeholder } from './components/placeholder';
 export { useDOMRect } from './hooks/use-dom-rect';
 export { VirtualGrid } from './components/virtual-grid';
